+++ src/providers/imdb/constants.ts
@@ -1,3 +1,4 @@
 export const IMDB_PROVIDER = "imdb";
 export const IMDB_TITLE_BASE_URL = "https://www.imdb.com/title";
 export const IMDB_DEFAULT_USER_AGENT = "@zivue/zuuid IMDb scraper (+https://github.com/zivue/zuuid)";
+export const IMDB_SUGGESTION_BASE_URL = "https://v3.sg.media-imdb.com/suggestion";