+++ src/providers/musicbrainz/index.ts
@@ -1,3 +1,4 @@
+export * from "./client.js";
 export * from "./artist.js";
 export * from "./constants.js";
 export * from "./label.js";
@@ -5,3 +6,4 @@ export * from "./recording.js";
 export * from "./release.js";
 export * from "./release-group.js";
 export * from "./work.js";
+export * from "./types.js";